+++ ui/packages/shared/profile/src/ProfileView/hooks/useResetStateOnSeriesChange.ts
@@ -12,31 +12,19 @@
 // limitations under the License.
 
 import {useURLState} from '@parca/components';
-import {useProfileFilters} from '../components/ProfileFilters/useProfileFilters';
 
-export const useResetStateOnProfileTypeChange = (): (() => void) => {
-  const [groupBy, setGroupBy] = useURLState('group_by');
-  const [searchString, setSearchString] = useURLState('search_string');
+export const useResetStateOnSeriesChange = (): (() => void) => {
   const [curPath, setCurPath] = useURLState('cur_path');
-  const {resetFilters} = useProfileFilters();
   const [sandwichFunctionName, setSandwichFunctionName] = useURLState('sandwich_function_name');
 
   return () => {
     setTimeout(() => {
-      if (groupBy !== undefined) {
-        setGroupBy(undefined);
-      }
-      if (searchString !== undefined) {
-        setSearchString(undefined);
-      }
       if (curPath !== undefined) {
         setCurPath(undefined);
       }
       if (sandwichFunctionName !== undefined) {
         setSandwichFunctionName(undefined);
       }
-
-      resetFilters();
     });
   };
 };